Update shared manifestwork methods
